beginning work on error codes
